1차 코드 리팩토링
