+++ src/basic/pages/cart/events/handleQuantityUpdate.js
@@ -1,9 +1,9 @@
 import { updateCartItemQuantity } from '../modules/updateCartItemQuantity.js';
-import { calcCart } from '../modules/calcCart.js';
+import { calculateCartTotals } from '../modules/calculateCartTotals.js';
 
 export function handleQuantityUpdate(event) {
   const target = event.target;
   const { itemId, change } = target.dataset;
   updateCartItemQuantity(itemId, parseInt(change));
-  calcCart(); // 장바구니 총액 계산
+  calculateCartTotals();
 }
