+++ src/basic/pages/cart/events/handleRemoveCartItem.js
@@ -1,5 +1,5 @@
 import { cartItemStore } from '../store.js';
-import { calcCart } from '../modules/calcCart.js';
+import { calculateCartTotals } from '../modules/calculateCartTotals.js';
 
 export function handleRemoveCartItem(event) {
   const target = event.target;
@@ -12,5 +12,5 @@ export function handleRemoveCartItem(event) {
     return { cartItems: updatedCartItems };
   });
 
-  calcCart(); // 장바구니 총액 계산
+  calculateCartTotals();
 }
