+++ src/basic/pages/cart/rerenders/rerenderStockStatus.js
@@ -1,4 +1,5 @@
-import { cartItemStore, globalCartStore } from '../store.js';
+import { cartItemStore } from '../store.js';
+import { DEFAULT_PRODUCT_LIST } from '../constant/defaultProducts.js';
 
 const createStockStatusText = (item) =>
   `${item.name}: ${item.quantity > 0 ? '재고 부족' + ` (${item.quantity}개 남음)` : '품절'}\n`;
@@ -7,11 +8,10 @@ export function rerenderStockStatus() {
   let updateContent = '';
   let currentContent = '';
 
-  const { productList } = globalCartStore.getState();
   const { cartItems } = cartItemStore.getState();
   const $stockInfoContainer = document.getElementById('stock-status');
 
-  const notQuantityList = productList.filter((product) => product.quantity === 0);
+  const notQuantityList = DEFAULT_PRODUCT_LIST.filter((product) => product.quantity === 0);
 
   if (notQuantityList.length > 0) {
     notQuantityList.forEach((item) => {
