+++ src/basic/pages/cart/store.js
@@ -1,15 +1,5 @@
 import { createStore } from '../../lib/createStore.js';
 
-export const globalCartStore = createStore({
-  productList: [
-    { id: 'p1', name: '상품1', price: 10000, quantity: 50 },
-    { id: 'p2', name: '상품2', price: 20000, quantity: 30 },
-    { id: 'p3', name: '상품3', price: 30000, quantity: 20 },
-    { id: 'p4', name: '상품4', price: 15000, quantity: 0 },
-    { id: 'p5', name: '상품5', price: 25000, quantity: 10 },
-  ],
-});
-
 export const cartItemStore = createStore({
   cartItems: [],
 });
@@ -17,9 +7,5 @@ export const cartItemStore = createStore({
 export const cartTotalPriceStore = createStore({
   totalPrice: 0,
   discountRate: 0,
-  rewardPoints: 0,
-});
-
-export const cartPointStore = createStore({
   rewardPoints: 0,
 });
